# openapi.v1.yaml
openapi: 3.1.1
info:
  title: TNPSC Book RAG API
  version: 1.0.0
  summary: Frontend contract for the English-first TNPSC textbook RAG MVP.
  description: |
    This checked-in contract contains implemented and planned v1 operations so the frontend can
    generate types and mocks before every backend route exists. `x-implementation-status` records
    current delivery state. FastAPI's generated OpenAPI must conform to the implemented subset.
servers:
  - url: "{origin}"
    variables:
      origin:
        default: http://127.0.0.1:8000
        description: Deployment-specific API origin.
...
paths:
...
  /v1/documents/{document_id}:
    parameters:
      - $ref: "#/components/parameters/DocumentId"
    get:
      tags: [catalog]
      operationId: getDocument
      summary: Get document detail and latest ingestion run.
      x-implementation-status: implemented
      security:
        - HTTPBearer: []
      parameters:
        - $ref: "#/components/parameters/RequestId"
      responses:
        "200":
          description: Document detail.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentDetail"
        "404":
          $ref: "#/components/responses/NotFound"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /v1/ingestion-runs:
    get:
      tags: [ingestion]
      operationId: listIngestionRuns
      summary: List ingestion operations across the catalog.
      x-implementation-status: implemented
      security:
        - HTTPBearer: []
      parameters:
        - $ref: "#/components/parameters/RequestId"
        - name: status
          in: query
          schema:
            type: array
            uniqueItems: true
            items:
              $ref: "#/components/schemas/IngestionRunStatus"
          style: form
          explode: true
        - name: stage
          in: query
          schema:
            type: array
            uniqueItems: true
            items:
              $ref: "#/components/schemas/IngestionStage"
          style: form
          explode: true
        - name: book_id
          in: query
          schema:
            type: string
            format: uuid
        - name: document_id
          in: query
          schema:
            type: string
            format: uuid
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/IncludeCount"
      responses:
        "200":
          description: Paginated global ingestion operations.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IngestionOperationPage"
        "422":
          $ref: "#/components/responses/Unprocessable"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /v1/documents/{document_id}/ingestion-runs:
    parameters:
      - $ref: "#/components/parameters/DocumentId"
    get:
      tags: [ingestion]
      operationId: listDocumentIngestionRuns
      summary: List ingestion runs for one document.
      x-implementation-status: implemented
      security:
        - HTTPBearer: []
      parameters:
        - $ref: "#/components/parameters/RequestId"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/IncludeCount"
      responses:
        "200":
          description: Paginated document ingestion runs.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IngestionRunPage"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/Unprocessable"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /v1/ingestion-runs/{run_id}:
    parameters:
      - $ref: "#/components/parameters/RunId"
    get:
      tags: [ingestion]
      operationId: getIngestionRun
      summary: Get an ingestion run and polling hint.
      x-implementation-status: implemented
      security:
        - HTTPBearer: []
      parameters:
        - $ref: "#/components/parameters/RequestId"
      responses:
        "200":
          description: Ingestion run state.
          headers:
            Cache-Control:
              schema:
                type: string
              example: no-store
            X-Request-ID:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IngestionRunDetailResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
...
  /v1/documents/{document_id}/pages:
    parameters:
      - $ref: "#/components/parameters/DocumentId"
    get:
      tags: [inspection]
      operationId: listDocumentPages
      summary: List extracted pages for a document.
      x-implementation-status: implemented
      security:
        - HTTPBearer: []
      parameters:
        - $ref: "#/components/parameters/RequestId"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/IncludeCount"
      responses:
        "200":
          description: Paginated page summaries.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageSummaryPage"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/Unprocessable"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /v1/pages/{page_id}:
    parameters:
      - $ref: "#/components/parameters/PageId"
    get:
      tags: [inspection]
      operationId: getPage
      summary: Get extraction detail for one page.
      x-implementation-status: implemented
      security:
        - HTTPBearer: []
      parameters:
        - $ref: "#/components/parameters/RequestId"
      responses:
        "200":
          description: Page detail.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageDetail"
        "404":
          $ref: "#/components/responses/NotFound"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
    patch:
      tags: [inspection]
      operationId: updatePrintedPageLabel
      summary: Correct the human-facing printed page label.
      x-implementation-status: implemented
      security:
        - HTTPBearer: []
      parameters:
        - $ref: "#/components/parameters/RequestId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePageRequest"
      responses:
        "200":
          description: Updated page detail.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageDetail"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/Unprocessable"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /v1/documents/{document_id}/chunks:
    parameters:
      - $ref: "#/components/parameters/DocumentId"
    get:
      tags: [inspection]
      operationId: listDocumentChunks
      summary: List extracted chunks for a document.
      x-implementation-status: implemented
      security:
        - HTTPBearer: []
      parameters:
        - $ref: "#/components/parameters/RequestId"
        - name: page_id
          in: query
          schema:
            type: string
            format: uuid
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/IncludeCount"
      responses:
        "200":
          description: Paginated chunk summaries.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChunkPage"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/Unprocessable"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
...
components:
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Short-lived access token issued by the configured OpenID Connect provider.
  parameters:
    RequestId:
      name: X-Request-ID
      in: header
      required: false
      schema:
        $ref: "#/components/schemas/RequestIdentifier"
...
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    Cursor:
      name: cursor
      in: query
      required: false
      schema:
        type: string
        minLength: 1
    IncludeCount:
      name: include_count
      in: query
      required: false
      schema:
        type: boolean
        default: false
...
    DocumentId:
      name: document_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    RunId:
      name: run_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    PageId:
      name: page_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  headers:
    RequestId:
      required: true
      schema:
        $ref: "#/components/schemas/RequestIdentifier"
...
  responses:
    NotFound:
      description: Resource was not found.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestId"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
    Unprocessable:
      description: Validation, cursor, or document-content error.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestId"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
    ServiceUnavailable:
      description: A required service is temporarily unavailable.
      headers:
        Retry-After:
          schema:
            oneOf:
              - type: integer
                minimum: 0
              - type: string
        X-Request-ID:
          $ref: "#/components/headers/RequestId"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
  schemas:
    RequestIdentifier:
      type: string
      minLength: 1
      maxLength: 200
      description: Accepted or generated safe printable correlation identifier.
    TextbookStandard:
      type: integer
      enum: [6, 7, 8, 9, 10]
...
    DocumentState:
      type: string
      enum: [uploaded, queued, extracting, chunking, embedding, ready, failed]
...
    IngestionRunStatus:
      type: string
      enum: [queued, running, succeeded, failed]
    IngestionStage:
      type: string
      enum: [queued, extraction, chunking, embedding, activation]
...
    ChunkContentType:
      type: string
      enum: [prose, heading, list, table, caption, mixed]
...
    ValidationFieldError:
      type: object
      required: [field, message, code]
      properties:
        field:
          type: string
          description: >-
            Stable path such as body.filters.standards.0, query.limit, path.book_id,
            header.Idempotency-Key, or form.file.
        message: {type: string}
        code: {type: string}
      additionalProperties: false
    Problem:
      type: object
      required: [type, title, status, detail, instance, code, request_id, errors]
      properties:
        type: {type: string, format: uri-reference}
        title: {type: string}
        status: {type: integer, minimum: 400, maximum: 599}
        detail: {type: string}
        instance: {type: string, format: uri-reference}
        code: {type: string}
        request_id:
          $ref: "#/components/schemas/RequestIdentifier"
        errors:
          type: array
          items:
            $ref: "#/components/schemas/ValidationFieldError"
      additionalProperties: false
    BookReference:
      type: object
      required: [id, title, standard, subject]
      properties:
        id: {type: string, format: uuid}
        title: {type: string}
        standard:
          $ref: "#/components/schemas/TextbookStandard"
        subject: {type: string}
...
    DocumentSummary:
      type: object
      required:
        - id
        - book_id
        - edition
        - source_filename
        - media_type
        - source_sha256
        - file_size_bytes
        - page_count
        - state
        - activated_at
        - created_at
        - updated_at
      properties:
        id: {type: string, format: uuid}
        book_id: {type: string, format: uuid}
        edition: {type: string, minLength: 1, maxLength: 200}
        source_filename: {type: string}
        media_type: {type: string, const: application/pdf}
        source_sha256: {type: string, pattern: "^[a-f0-9]{64}$"}
        file_size_bytes: {type: integer, minimum: 1}
        page_count: {type: [integer, "null"], minimum: 1}
        state:
          $ref: "#/components/schemas/DocumentState"
        activated_at: {type: [string, "null"], format: date-time}
        created_at: {type: string, format: date-time}
        updated_at: {type: string, format: date-time}
...
    DocumentReference:
      type: object
      required: [id, edition, source_filename, state]
      properties:
        id: {type: string, format: uuid}
        edition: {type: string}
        source_filename: {type: string}
        state:
          $ref: "#/components/schemas/DocumentState"
    IngestionIssue:
      type: object
      required: [code, message, stage, pdf_page_index]
      properties:
        code: {type: string}
        message: {type: string}
        stage:
          oneOf:
            - $ref: "#/components/schemas/IngestionStage"
            - type: "null"
        pdf_page_index: {type: [integer, "null"], minimum: 0}
      additionalProperties: false
    IngestionRun:
      type: object
      required:
        - id
        - document_id
        - status
        - current_stage
        - retry_count
        - started_at
        - completed_at
        - warnings
        - error
        - created_at
        - updated_at
      properties:
        id: {type: string, format: uuid}
        document_id: {type: string, format: uuid}
        status:
          $ref: "#/components/schemas/IngestionRunStatus"
        current_stage:
          $ref: "#/components/schemas/IngestionStage"
        retry_count: {type: integer, minimum: 0}
        started_at: {type: [string, "null"], format: date-time}
        completed_at: {type: [string, "null"], format: date-time}
        warnings:
          type: array
          items:
            $ref: "#/components/schemas/IngestionIssue"
        error:
          oneOf:
            - $ref: "#/components/schemas/IngestionIssue"
            - type: "null"
        created_at: {type: string, format: date-time}
        updated_at: {type: string, format: date-time}
      description: >-
        Valid status/stage/timestamp/error combinations are normative in api_spec.md. A retry
        preserves warnings, increments retry_count, keeps the failed stage, and clears execution
        timestamps and error before re-entering the queue. The service does not expose a synthetic
        progress percentage.
    IngestionRunDetailResponse:
      type: object
      required: [ingestion_run, poll_after_seconds]
      properties:
        ingestion_run:
          $ref: "#/components/schemas/IngestionRun"
        poll_after_seconds: {type: integer, minimum: 1}
    IngestionOperationItem:
      type: object
      required: [ingestion_run, document, book]
      properties:
        ingestion_run:
          $ref: "#/components/schemas/IngestionRun"
        document:
          $ref: "#/components/schemas/DocumentReference"
        book:
          $ref: "#/components/schemas/BookReference"
    DocumentDetail:
      allOf:
        - $ref: "#/components/schemas/DocumentSummary"
        - type: object
          required: [latest_ingestion_run]
          properties:
            latest_ingestion_run:
              oneOf:
                - $ref: "#/components/schemas/IngestionRun"
                - type: "null"
    PaginationMetadata:
      type: object
      required: [previous_cursor, next_cursor, total_items]
      properties:
        previous_cursor: {type: [string, "null"]}
        next_cursor: {type: [string, "null"]}
        total_items: {type: [integer, "null"], minimum: 0}
      description: >-
        total_items is non-null only when include_count=true. Cursors support adjacent navigation;
        arbitrary page-number jumps are not supported.
...
    IngestionOperationPage:
      allOf:
        - $ref: "#/components/schemas/PaginationMetadata"
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/IngestionOperationItem"
    IngestionRunPage:
      allOf:
        - $ref: "#/components/schemas/PaginationMetadata"
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/IngestionRun"
...
    PageSummary:
      type: object
      required:
        - id
        - document_id
        - pdf_page_index
        - printed_page_label
        - width
        - height
        - warning_count
        - created_at
      properties:
        id: {type: string, format: uuid}
        document_id: {type: string, format: uuid}
        pdf_page_index: {type: integer, minimum: 0}
        printed_page_label: {type: [string, "null"], maxLength: 100}
        width: {type: [number, "null"], exclusiveMinimum: 0}
        height: {type: [number, "null"], exclusiveMinimum: 0}
        warning_count: {type: integer, minimum: 0}
        created_at: {type: string, format: date-time}
    PageSummaryPage:
      allOf:
        - $ref: "#/components/schemas/PaginationMetadata"
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/PageSummary"
    UpdatePageRequest:
      type: object
      required: [printed_page_label]
      properties:
        printed_page_label: {type: [string, "null"], maxLength: 100}
      additionalProperties: false
...
    Asset:
      allOf:
        - $ref: "#/components/schemas/AssetReference"
        - type: object
          required: [page_id, mime_type, sha256, bounding_box, created_at]
          properties:
            page_id: {type: string, format: uuid}
            mime_type: {type: string}
            sha256: {type: string, pattern: "^[a-f0-9]{64}$"}
            bounding_box:
              oneOf:
                - $ref: "#/components/schemas/BoundingBox"
                - type: "null"
            created_at: {type: string, format: date-time}
    ChunkSummary:
      type: object
      required:
        - id
        - page_id
        - document_id
        - sequence_number
        - display_text
        - chapter_title
        - section_path
        - content_type
        - token_count
        - created_at
      properties:
        id: {type: string, format: uuid}
        page_id: {type: string, format: uuid}
        document_id: {type: string, format: uuid}
        sequence_number: {type: integer, minimum: 0}
        display_text: {type: string}
        chapter_title: {type: [string, "null"]}
        section_path:
          type: array
          items: {type: string}
        content_type:
          $ref: "#/components/schemas/ChunkContentType"
        token_count: {type: integer, minimum: 0}
        created_at: {type: string, format: date-time}
    ChunkPage:
      allOf:
        - $ref: "#/components/schemas/PaginationMetadata"
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/ChunkSummary"
    PageDetail:
      allOf:
        - $ref: "#/components/schemas/PageSummary"
        - type: object
          required: [raw_text, normalized_text, warnings, chunks, assets]
          properties:
            raw_text: {type: string}
            normalized_text: {type: string}
            warnings:
              type: array
              items:
                $ref: "#/components/schemas/IngestionIssue"
            chunks:
              type: array
              items:
                $ref: "#/components/schemas/ChunkSummary"
            assets:
              type: array
              items:
                $ref: "#/components/schemas/Asset"
...